requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
icalendar>=5.0.13
python-dateutil>=2.9.0.post0
beautifulsoup4>=4.12.3
tzdata>=2024.1
